restructure to use http package
